add task 9, 11, 13. done some improvements and changes in code

+++ Tasks/Task3.java
@@ -2,16 +2,16 @@ import java.util.Scanner;
 
 public class Task3 {
 	static String input;
-	static int rate;
+	static double rate;
 	static int amouth;
 	static int dollar;
-	static int change;
+	static double change;
 
 	public void task3(Scanner sc) {
 		System.out.println("Input cost of dollar and sum of money in gryvna:");
-		rate = sc.nextInt();
+		rate = sc.nextDouble();
 		amouth = sc.nextInt();
-		dollar = amouth / rate;
+		dollar = (int) (amouth / rate);
 		change = amouth % rate;
 		System.out.println("You can buy " + dollar + " dollars with " + change + " in change");
 	}
